Improve types for onChangeProp
