+++ packages/toolpad-app/src/toolpad/AppEditor/PageEditor/NodeAttributeEditor.tsx
@@ -19,11 +19,11 @@ import MarkdownTooltip from '../../../components/MarkdownTooltip';
 import { isTemplateDescendant } from '../../../toolpadComponents/template';
 import { NON_BINDABLE_CONTROL_TYPES } from '../../../constants';
 
-export interface NodeAttributeEditorProps<P extends object> {
+export interface NodeAttributeEditorProps<P extends object, K extends keyof P = keyof P> {
   node: appDom.AppDomNode;
   namespace?: string;
   name: string;
-  argType: ArgTypeDefinition<P>;
+  argType: ArgTypeDefinition<P, K>;
   props?: P;
 }
