+++ packages/toolpad-core/src/types.ts
@@ -165,11 +165,24 @@ export type PropValueType =
   | TemplateValueType
   | EventValueType;
 
+interface ParameterTypeLookup {
+  number: number;
+  string: string;
+  boolean: boolean;
+  array: unknown[];
+  object: Record<string, unknown>;
+  element: React.ReactNode;
+  template: () => React.ReactNode;
+  event: (...args: any[]) => void;
+}
+
+export type InferParameterType<T extends PropValueType> = ParameterTypeLookup[T['type']];
+
 export type PropValueTypes<K extends string = string> = Partial<{
   [key in K]?: PropValueType;
 }>;
 
-export interface ArgTypeDefinition<P extends object = {}, V = P[keyof P]> {
+export interface ArgTypeDefinition<P, K extends keyof P = keyof P> {
   /**
    * A short explanatory text that'll be shown in the editor UI when this property is referenced.
    * May contain Markdown.
@@ -195,21 +208,21 @@ export interface ArgTypeDefinition<P extends object = {}, V = P[keyof P]> {
    * A default value for the property.
    * @deprecated Use `typeDef.default` instead.
    */
-  defaultValue?: V;
+  defaultValue?: P[K];
   /**
    * The property that will supply the default value.
    */
-  defaultValueProp?: V;
+  defaultValueProp?: keyof P & string;
   /**
    * The property that is used to control this property.
    */
-  onChangeProp?: string;
+  onChangeProp?: keyof P & string;
   /**
    * Provides a way to manipulate the value from the onChange event before it is assigned to state.
    * @param {...any} params params for the function assigned to [onChangeProp]
    * @returns {any} a value for the controlled prop
    */
-  onChangeHandler?: (...params: any[]) => V;
+  onChangeHandler?: (...params: any[]) => P[K];
   /**
    * For compound components, this property is used to control the visibility of this property based on the selected value of another property.
    * If this property is not defined, the property will be visible at all times.
@@ -225,12 +238,11 @@ export interface ArgTypeDefinition<P extends object = {}, V = P[keyof P]> {
 }
 
 export type ArgTypeDefinitions<P extends object = {}> = {
-  [K in keyof P & string]?: ArgTypeDefinition<P, P[K]>;
+  [K in keyof P & string]?: ArgTypeDefinition<P, K>;
 };
 
 export interface ComponentDefinition<P extends object = {}> {
-  // props: PropDefinitions<P>;
-  argTypes: ArgTypeDefinitions<P>;
+  argTypes?: ArgTypeDefinitions<P>;
 }
 
 export interface LiveBindingError {
